FIX: viewmodel 삭제, appComponent에서 repository 주입
