+++ MyLibrary/View/HomeViewController/HomeViewController.swift
@@ -14,6 +14,7 @@ import RxDataSources
 
 import Model
 import ViewModel
+import Repository
 
 public protocol HomeRouting {
     func addMemoViewController(messageHandler: @escaping (String) -> ()) -> UIViewController
@@ -27,7 +28,7 @@ public class HomeViewController: UICollectionViewController {
     private let reactor: Reactor
     private var disposeBag: DisposeBag = .init()
     private let routing: HomeRouting
-    private let memoViewModel = MemoViewModel()
+    private let memoRepository: MemoRepository
     
     typealias DataSource = RxCollectionViewSectionedReloadDataSource<HomeSection>
     lazy var dataSource = DataSource(configureCell: { dataSource, collectionView, indexPath, item in
@@ -71,9 +72,10 @@ public class HomeViewController: UICollectionViewController {
         }
     })
     
-    public init(reactor: HomeViewReactor, routing: HomeRouting) {
+    public init(reactor: HomeViewReactor, routing: HomeRouting, memoRepository: MemoRepository) {
         self.reactor = reactor
         self.routing = routing
+        self.memoRepository = memoRepository
         let layout = UICollectionViewFlowLayout()
         super.init(collectionViewLayout: layout)
     }
@@ -129,9 +131,10 @@ public class HomeViewController: UICollectionViewController {
             .subscribe(onNext: { [weak self ] indexPath in
                 guard let self = self else { return }
                 let viewController = self.routing.addMemoViewController { [weak self] memo in
-                    if self?.memoViewModel.create(content: memo) ?? false {
+                    if self?.memoRepository.create(content: memo) ?? false {
                         // 메모 create 성공
                         print("Success insert into Memo: \(memo)")
+                        // self?.memoRepository.fetch()
                     }
                 }
                 self.present(viewController, animated: true)
