+++ MyLibrary/ViewModel/HomeViewReactor.swift
@@ -42,11 +42,14 @@ public class HomeViewReactor: Reactor {
     
     private let todoRepository: TodoRepository
     private let planRepository: PlanRepository
+    private let memoRepository: MemoRepository
+    
     private var sections = [HomeSection]()
     
-    public init(todoRepository: TodoRepository, planRepository: PlanRepository) {
+    public init(todoRepository: TodoRepository, planRepository: PlanRepository, memoRepository: MemoRepository) {
         self.todoRepository = todoRepository
         self.planRepository = planRepository
+        self.memoRepository = memoRepository
         
         self.initialState = State(
             sections: [HomeSection]()
